mboxlist: factor jmapid repair out of ctl_mboxlist for reuse
ctl_mboxlist -k already knew how to assign a jmapid and write the J
record.  The audit needs exactly that, and a second implementation would
drift from the first -- which is the problem this whole port exists to
solve -- so the body of fix_cb moves to mboxlist_fix_jmapid() and
ctl_mboxlist calls it.

Two leaks fixed while moving it: byunqid was leaked on the
already-has-a-record early return, and the mailbox opened to assign a new
jmapid was never closed.

The audit applies these outside the delete transaction, since the repair
opens the mailbox and bumps its modseq, and takes the user nslock.  Each
mailbox is looked up again first: the sweep produced a candidate list, not
a decision.
