+++ imap/mboxlist.c
@@ -31,6 +31,7 @@
 #include "cyrusdb.h"
 #include "hash.h"
 #include "util.h"
+#include "jmap_util.h"
 #include "mailbox.h"
 #include "mboxevent.h"
 #include "xmalloc.h"
@@ -1246,6 +1247,91 @@ EXPORTED int mboxlist_foreach_raw(mboxlist_rawproc_t *proc, void *rock,
     /* One foreach inside a single transaction gives a consistent snapshot
      * across all key types, which every cross-check depends on. */
     return cyrusdb_foreach(mbdb, "", 0, NULL, rawforeach_cb, &frock, tid);
+}
+
+/*
+ * Ensure this mailbox is reachable by jmapid: assign one if it has none,
+ * and write the J record.
+ *
+ * This is not a database-only repair.  The jmapid is derived from the
+ * folder's modseq, so with no jmapid present it opens the mailbox,
+ * dirties the modseq, writes the header and commits before updating the
+ * mbentry.  Callers must be prepared for that cost.
+ *
+ * Shared by ctl_mboxlist -k and the consistency audit: a second
+ * implementation of jmapid assignment would drift from this one.
+ */
+EXPORTED int mboxlist_fix_jmapid(const mbentry_t *mbentry)
+{
+    mbentry_t *byunqid = NULL;
+
+    int r = mboxlist_lookup_by_uniqueid(mbentry->uniqueid, &byunqid, NULL);
+    if (r) {
+        xsyslog(LOG_NOTICE, "missing uniqueid record, skipping",
+               "mboxname=<%s> uniqueid=<%s>",
+               mbentry->name, mbentry->uniqueid);
+        return 0;
+    }
+
+    // we are not the current record?  We don't need to process this.
+    if (strcmp(mbentry->name, byunqid->name))
+        goto done;
+
+    if (mbentry->jmapid) {
+        mbname_t *mbname = mbname_from_intname(mbentry->name);
+        const char *userid = mbname_userid(mbname);
+
+        if (!userid) userid = "";
+
+        // already got a record, we're good!
+        int res = mboxlist_lookup_by_jmapid(userid, mbentry->jmapid, NULL, NULL);
+        mbname_free(&mbname);
+        if (!res) goto done;
+
+        xsyslog(LOG_NOTICE, "adding missing J record to mboxlist",
+                "mboxname=<%s> jmapid=<%s>",
+                mbentry->name, mbentry->jmapid);
+    }
+    else {
+        struct mailbox *mailbox = NULL;
+        r = mailbox_open_from_mbe(byunqid, &mailbox);
+        if (r) goto done;
+        struct buf jmapid = BUF_INITIALIZER;
+
+        byunqid->foldermodseq = mailbox_modseq_dirty(mailbox);
+
+        buf_putc(&jmapid, JMAP_MAILBOXID_PREFIX);
+        MODSEQ_TO_JMAPID(&jmapid, byunqid->foldermodseq);
+        mailbox_set_jmapid(mailbox, buf_cstring(&jmapid));
+
+        free(byunqid->jmapid);
+        byunqid->jmapid = buf_release(&jmapid);
+        buf_free(&jmapid);
+
+        xsyslog(LOG_NOTICE, "adding new J record to mboxlist",
+                "mboxname=<%s> jmapid=<%s>",
+                byunqid->name, byunqid->jmapid);
+
+        r = mailbox_commit(mailbox);
+        mailbox_close(&mailbox);
+        if (r) {
+            xsyslog(LOG_ERR, "DBERROR: error committing transaction",
+                    "error=<%s>", cyrusdb_strerror(r));
+            goto done;
+        }
+    }
+
+    r = mboxlist_updatelock(byunqid, /*localonly*/1);
+    if (r) {
+        xsyslog(LOG_ERR, "failed to update mboxlist",
+                "mboxname=<%s> error=<%s>",
+                mbentry->name, error_message(r));
+    }
+
+done:
+    mboxlist_entry_free(&byunqid);
+
+    return r;
 }
 
 EXPORTED void mboxlist_key_for_name(const char *mboxname, struct buf *key)
